refactor(web-inspector): extract thread inspection domain
