+++ packages/web-inspector/src/__tests__/web-inspector.spec.ts
@@ -552,22 +552,6 @@ type ThreadDetailsInternals = {
   renderConversation: () => unknown;
   renderState: () => unknown;
   renderEvents: () => unknown;
-  timelineItemsFromEvents: (
-    events: Array<Record<string, unknown>>,
-  ) => Array<Record<string, unknown>>;
-  mapMessages: (
-    messages: Array<{
-      id: string;
-      role: string;
-      content?: string | null;
-      toolCallId?: string | null;
-      toolCalls?: Array<{
-        id: string;
-        name: string;
-        args: string;
-      }>;
-    }>,
-  ) => Array<Record<string, unknown>>;
 };
 
 function createThreadDetails(): {
@@ -838,12 +822,11 @@ describe("ɵCpkThreadDetails caching", () => {
       },
     ];
 
-    const normalizeSpy = vi.spyOn(internals, "timelineItemsFromEvents");
-
     const timelineTpl = internals.renderTimeline();
+    const timelineItems = internals._timelineItemsCache?.items;
 
     expect(internals.renderTimeline()).toBe(timelineTpl);
-    expect(normalizeSpy).toHaveBeenCalledTimes(1);
+    expect(internals._timelineItemsCache?.items).toBe(timelineItems);
   });
 
   it("keeps live fallback event references stable for timeline and raw event caches", async () => {
@@ -858,14 +841,13 @@ describe("ɵCpkThreadDetails caching", () => {
       },
     ];
 
-    const normalizeSpy = vi.spyOn(internals, "timelineItemsFromEvents");
-
     const timelineTpl = internals.renderTimeline();
+    const timelineItems = internals._timelineItemsCache?.items;
     const eventsTpl = internals.renderEvents();
 
     expect(internals.renderTimeline()).toBe(timelineTpl);
     expect(internals.renderEvents()).toBe(eventsTpl);
-    expect(normalizeSpy).toHaveBeenCalledTimes(1);
+    expect(internals._timelineItemsCache?.items).toBe(timelineItems);
   });
 
   it("state and events caches invalidate when their fetched data is reassigned", async () => {
@@ -900,57 +882,6 @@ describe("ɵCpkThreadDetails caching", () => {
 
     expect(internals.renderState()).not.toBe(stateA);
     expect(internals.renderEvents()).not.toBe(eventsA);
-  });
-
-  it("maps empty tool arguments and results as empty objects", () => {
-    const { internals } = createThreadDetails();
-
-    const items = internals.mapMessages([
-      {
-        id: "assistant-1",
-        role: "assistant",
-        content: "",
-        toolCalls: [
-          {
-            id: "call-empty",
-            name: "lookupUser",
-            args: "",
-          },
-          {
-            id: "call-whitespace",
-            name: "lookupOrder",
-            args: "   ",
-          },
-        ],
-      },
-      {
-        id: "tool-empty",
-        role: "tool",
-        toolCallId: "call-empty",
-        content: "",
-      },
-      {
-        id: "tool-whitespace",
-        role: "tool",
-        toolCallId: "call-whitespace",
-        content: "   ",
-      },
-    ]);
-
-    expect(items).toMatchObject([
-      {
-        type: "tool_call",
-        toolCallId: "call-empty",
-        arguments: {},
-        result: {},
-      },
-      {
-        type: "tool_call",
-        toolCallId: "call-whitespace",
-        arguments: {},
-        result: {},
-      },
-    ]);
   });
 });
